requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
itchat==1.3.10
Pillow==5.3.0
pypng==0.0.18
PyQRCode==1.2.1
requests==2.21.0
urllib3==1.24.1
